server/requirements.txt
Pillow
psycopg2
gunicorn
Django
django-debug-toolbar
djangorestframework
django-cors-headers
djoser
drf_spectacular
django-filter
requests
celery
redis
django-celery-results
django-celery-beat
geoip2
